% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-meta.R
\name{get_meta}
\alias{get_meta}
\title{Extract Series meta information.}
\usage{
get_meta(series_id = seriespicker::ids)
}
\arguments{
\item{series_id}{Character string of the IMDB series id.}
}
\value{
Character giving the original series title.
}
\description{
Extract Series meta information.
}
\examples{
series_id <- seriespicker::ids[c(1, 4, 6)]
get_meta(series_id)

get_meta(ids[1:10])

get_meta()
}
