% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rating.R
\name{get_rating}
\alias{get_rating}
\title{Get episode specific series rating.}
\usage{
get_rating(id, season)
}
\arguments{
\item{id}{Character IMDB series id string. E.g. \code{0944947} for "Game of Thrones".}

\item{season}{Integer Season number.}
}
\value{
tidy dataframe with columns "gender", "age", "rating", "votes", "id", "name"
}
\description{
Get episode specific series rating.
}
\examples{
id <- "0944947"
season <- 3
season <- 1

id <- "5491994"
season <- 1

id <- "1641384"
season <- 3

rating(id, season)

df <- rating("0944947", 5)
df <- rating("0052520", 3)
}
