% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-seasons.R
\name{get_seasons}
\alias{get_seasons}
\title{Extract number of seasons}
\usage{
get_seasons(series_id = seriespicker::ids)
}
\arguments{
\item{series_id}{Character string of the IMDB series id.}
}
\value{
Integer giving the number of seasons per series.
}
\description{
Extract number of seasons
}
\examples{
get_seasons(ids[1:10])
}
